fix: CI workflow and clippy warnings
- Fix actions/checkout@v6 → v4 (v6 doesn't exist)
- Fix MSRV check: 1.85 → 1.89 (matches Cargo.toml rust-version)
- Fix clippy: suppress dead_code on reserved constants, use div_ceil(),
  remove redundant i32 casts in strategy_bench

+++ examples/strategy_bench.rs
@@ -42,9 +42,9 @@ fn make_screenshot(width: usize, height: usize) -> Vec<u8> {
                 (next() % 6) as i32 - 3
             };
 
-            data.push((r as i32 + noise).clamp(0, 255) as u8);
-            data.push((g as i32 + noise).clamp(0, 255) as u8);
-            data.push((b as i32 + noise).clamp(0, 255) as u8);
+            data.push((r + noise).clamp(0, 255) as u8);
+            data.push((g + noise).clamp(0, 255) as u8);
+            data.push((b + noise).clamp(0, 255) as u8);
             data.push(255);
         }
     }
